Move data formatting for registrations to separate file for use in multiple places
